Feat/(#14): 동행 게시글 리스트 컴포넌트 파라미터 적용 및 스타일 구조 개선
