+++ src/app/page.tsx
@@ -1,9 +1,3 @@
-import TravelPostList from '@/components/board/boardlist'
-
 export default async function Home() {
-  return (
-    <div>
-      <TravelPostList />
-    </div>
-  )
+  return <div>프로젝트초기화</div>
 }
