+++ src/components/board/TravelPostCard/index.tsx
@@ -1,0 +1,92 @@
+import { Heart, User } from 'lucide-react'
+import Image from 'next/image'
+
+import { TravelPost } from '@/types/meeting/travel'
+
+export default function TravelPostCard({ post }: { post: TravelPost }) {
+  const tagStyle = 'px-3 py-1 bg-blue-50 text-main rounded-full text-xs'
+  const cardBase =
+    'bg-white rounded-2xl p-6 shadow-sm hover:shadow-md transition-shadow border border-border'
+
+  return (
+    <div className={cardBase}>
+      <div className="flex gap-4">
+        <div className="relative w-[188px] h-[188px] rounded-2xl overflow-hidden shrink-0 bg-bg-disabled">
+          <Image src={post.thumbnail} alt="썸네일" fill />
+        </div>
+
+        {/* Content */}
+        <div className="flex-1 ml-3">
+          <div className="flex gap-2 mb-2">
+            <span className={tagStyle}>여행스타일 태그</span>
+            <span className={tagStyle}>숙소취향 태그</span>
+          </div>
+
+          <h3 className="text-lg font-semibold text-text-base mb-1">
+            {post.title}
+          </h3>
+
+          <div className="flex gap-1">
+            <p className="text-sm text-text-disabled mb-3">작성자</p>
+            <p className="text-sm text-text-input">작성자 닉네임</p>
+          </div>
+
+          {/* 신청자 */}
+          <div className="flex text-sm mb-2 gap-1 mt-8">
+            <User className="w-4 h-4" />
+            <span>{post.currentMembers}명 신청</span>
+          </div>
+
+          {/* 기타 정보 */}
+          <div className="flex items-center gap-2 text-sm text-text-input">
+            <div className="flex items-center gap-1">
+              <span className="text-text-disabled">위치</span>
+              <span className="text-text-base">{post.region}</span>
+            </div>
+
+            <span className="text-text-disabled">|</span>
+
+            <div className="flex items-center gap-1">
+              <span className="text-text-disabled">날짜</span>
+              <span className="text-text-base">
+                {new Date(post.period.startDate).toLocaleDateString('ko-KR', {
+                  month: 'long',
+                  day: 'numeric',
+                })}
+              </span>
+            </div>
+
+            <span className="text-text-disabled">|</span>
+
+            <div className="flex items-center gap-1">
+              <span className="text-text-disabled">나이</span>
+              <span className="text-text-base">20대만</span>
+            </div>
+
+            <span className="text-text-disabled">|</span>
+
+            <div className="flex items-center gap-1">
+              <span className="text-text-disabled">성별</span>
+              <span className="text-text-base">남여자만</span>
+            </div>
+          </div>
+        </div>
+
+        <div className="flex flex-col items-end justify-between">
+          <button className="w-10 h-10 flex items-center justify-center hover:bg-bg-input rounded-full transition-colors">
+            <Heart
+              className={`w-6 h-6 ${
+                post.bookmarked ? 'fill-main text-main' : 'text-text-input'
+              }`}
+            />
+          </button>
+
+          {/* 버튼 */}
+          <button className="px-6 py-2 rounded-lg font-medium transition-colors bg-main text-white hover:opacity-90">
+            신청하기
+          </button>
+        </div>
+      </div>
+    </div>
+  )
+}
