+++ src/types/meeting/travel.ts
@@ -1,0 +1,47 @@
+// src/types/travel.ts
+
+export type TravelSearchParams = {
+  region?: string
+  startDate?: string
+  endDate?: string
+  status?: 'RECRUITING' | 'CLOSED'
+  ageRange?: string
+  gender?: 'MALE' | 'FEMALE'
+  keyword?: string
+  page?: number
+  size?: number
+}
+
+export type TravelPeriod = {
+  startDate: string
+  endDate: string
+}
+
+export type TravelPost = {
+  postId: string
+  title: string
+  region: string
+  period: TravelPeriod
+  status: 'RECRUITING' | 'CLOSED'
+  currentMembers: number
+  maxMembers: number
+  bookmarked: boolean | null
+  thumbnail: string
+}
+
+export type PageInfo = {
+  page: number
+  size: number
+  totalPages: number
+  totalElements: number
+}
+
+export type TravelPostListResponse = {
+  success: boolean
+  status: number
+  timestamp: string
+  data: {
+    posts: TravelPost[]
+    pageInfo: PageInfo
+  }
+}
